requirements.txt
# requirements.txt

ftfy
ml-starter
ml-pretrained
numpy
tokenizers
torch>=2.0.0
